+++ cleanmarl/maddpg_multienvs.py
@@ -57,7 +57,7 @@ class Args:
     """ Update the target network each target_network_update_freq» step in the environment"""
     polyak: float = 0.01
     """ Polyak coefficient when using polyak averaging for target network update"""
-    clip_gradients: int = -1
+    clip_gradients: float = -1
     """ 0< for no clipping and 0> if clipping at clip_gradients"""
     log_every: int = 10
     """ Logging steps """
